requirements.txt
anyio==4.11.0
asgiref==3.10.0
astroid==3.3.8
autopep8==2.3.2
babel==2.17.0
black==25.1.0
Brotli==1.1.0
CacheControl==0.14.3
cachetools==6.2.1
certifi==2024.8.30
cffi==2.0.0
cfgv==3.4.0
channels==4.3.1
channels_redis==4.3.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
crispy-bootstrap5==2024.10
cryptography==46.0.2
cssselect2==0.8.0
diff-match-patch==20241021
dill==0.3.9
distlib==0.3.9
Django==4.2
django-admin-interface==0.29.1
django-appconf==1.0.6
django-background-tasks==1.2.8
django-colorfield==0.11.0
django-compressor==4.5.1
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==25.1
django-import-export==4.3.7
django-js-asset==3.1.2
django-money==3.5.3
django-mptt==0.16.0
django-multiselectfield==0.1.13
django-registration-redux==2.13
django-simple-history==3.8.0
django-tables2==2.7.0
django-tinymce==4.1.0
django-user-sessions==2.0.0
djangorestframework==3.16.1
easy-thumbnails==2.10
et_xmlfile==2.0.0
fcm-django==2.3.1
filelock==3.17.0
firebase_admin==7.1.0
flake8==7.1.2
fonttools==4.59.1
git-filter-repo==2.47.0
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
identify==2.6.8
idna==3.10
isort==6.0.1
mccabe==0.7.0
msgpack==1.1.2
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
openpyxl==3.1.5
packaging==24.2
pathspec==0.12.1
pdfkit==1.0.0
pdfrw==0.4
pillow==11.1.0
platformdirs==4.3.6
pre_commit==4.1.0
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
py-moneyed==3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
pydyf==0.11.0
pyflakes==3.2.0
PyJWT==2.10.1
pylint==3.3.4
PyPDF2==3.0.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==8.2
razorpay==2.0.0
rcssmin==1.1.2
redis==6.4.0
reportlab==4.3.1
requests==2.32.5
rjsmin==1.2.2
rsa==4.9.1
setuptools==80.9.0
sib-api-v3-sdk==7.6.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
swapper==1.4.0
tablib==3.8.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.5.0
virtualenv==20.29.2
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
